Disable motion detection by default on desktop browsers
This isn't perfect, but it changes how Plax works by default on desktop.
Previously it would always be enabled if the browser supported it
(Firefox only?). Now it will be disabled unless you explicitly allow
desktop support. Detection is done by detecting 'ontouchstart' which is
only on mobile (for now).